extract vacationsFlow into reusable usecase
