+++ app/src/main/java/rocks/poopjournal/vacationdays/presentation/screen/add/AddViewModel.kt
@@ -1,7 +1,5 @@
 package rocks.poopjournal.vacationdays.presentation.screen.add
 
-import android.content.Context
-import android.widget.Toast
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -9,12 +7,14 @@ import kotlinx.coroutines.launch
 import rocks.poopjournal.vacationdays.data.VacationData
 import rocks.poopjournal.vacationdays.domain.repo.VacationRepository
 import rocks.poopjournal.vacationdays.presentation.ui.utils.ThemeSetting
+import rocks.poopjournal.vacationdays.presentation.usecase.VacationDataUseCase
 import javax.inject.Inject
 
 @HiltViewModel
 class AddViewModel @Inject constructor(
     private val vacationRepository: VacationRepository,
-    val themeSetting: ThemeSetting
+    val themeSetting: ThemeSetting,
+    val vacationsUseCase: VacationDataUseCase,
 ) : ViewModel() {
